+++ packages/spruce-skill-booter/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [29.1.4](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.1.3...v29.1.4) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [29.1.3](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.1.2...v29.1.3) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [29.1.2](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.1.1...v29.1.2) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [29.1.1](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.1.0...v29.1.1) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+# [29.1.0](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.0.11...v29.1.0) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [29.0.11](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.0.10...v29.0.11) (2022-03-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
 ## [29.0.10](https://github.com/sprucelabsai/spruce-features-workspace/compare/v29.0.9...v29.0.10) (2022-03-09)
 
 **Note:** Version bump only for package @sprucelabs/spruce-skill-booter